+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 * **@dlr-eoc/map-ol:**
   - Removed `MapOlService.layers_on_click(evt)` and `MapOlService.layers_on_pointermove(evt)` -> replace them by `MapOlService.layersOnMapEvent(evt)`.
   - Renamed `MapOlService.layers_on_click_move(evt, layerFilter)` -> `MapOlService.layersOnMapEvent(evt)` layer filtering is now done later.
+  - Renamed `MapOlService.layer_on_click(evt, layer, color?)` -> `MapOlService.layerOnEvent(evt, layer, color?)`.
+  - Renamed `MapOlService.raster_on_click(evt, layer, color?)` -> `MapOlService.rasterOnEvent(evt, layer, color?)`.
+  - Renamed `MapOlService.vector_on_click(evt)` -> `MapOlService.vectorOnEvent(evt)`.
 
 * **@dlr-eoc/services-layers:**
   - Removed zIndex from types `ILayerOptions` which was forgotten to do at breaking version `6.0.0`.